[M8-25] [Back-End REST] Implement authentication. Partially implemented security feature.

+++ sma-site-server/src/main/java/org/fabasoad/rest/BaseResource.java
@@ -4,10 +4,13 @@ import org.fabasoad.db.dao.BaseDao;
 import org.fabasoad.db.dao.DaoFactory;
 import org.fabasoad.db.dao.DaoType;
 import org.fabasoad.db.pojo.BasePojo;
+import org.glassfish.jersey.server.ResourceConfig;
+import org.glassfish.jersey.server.filter.RolesAllowedDynamicFeature;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -25,7 +28,7 @@ import static org.fabasoad.api.Logger.getLogger;
  * @author efabizhevsky
  * @date 11/22/2016.
  */
-abstract class BaseResource {
+abstract class BaseResource extends ResourceConfig {
 
     Path getUploadPath() {
         return null;
@@ -38,6 +41,10 @@ abstract class BaseResource {
     abstract Function<String, Optional<String>> fromDto();
 
     abstract Map<String, String> getPojoProperties();
+
+    BaseResource() {
+        register(RolesAllowedDynamicFeature.class);
+    }
 
     Response getAll() {
         BaseDao<? extends BasePojo> dao = DaoFactory.create(getDaoType());
@@ -58,29 +65,41 @@ abstract class BaseResource {
         return Response.ok(json.toJSONString()).build();
     }
 
-    Response delete(int id) {
-        BaseDao<? extends BasePojo> dao = DaoFactory.create(getDaoType());
-        dao.delete(id);
-        String message = "Entity with id = " + id + " deleted successfully";
-        return Response.ok(buildSuccess(message).toJSONString()).build();
-    }
-
-    void upload(InputStream fileInputStream, String fileName) {
-        try (OutputStream out = new FileOutputStream(new File(getUploadPath().toString(), fileName))) {
-            int read;
-            byte[] bytes = new byte[1024];
-
-            while ((read = fileInputStream.read(bytes)) != -1) {
-                out.write(bytes, 0, read);
-            }
-        } catch (IOException e) {
-            getLogger().error(getClass(), String.format("Error while uploading '%s' file", fileName));
+    Response delete(SecurityContext context, int id) {
+        if (context.isUserInRole("admin")) {
+            BaseDao<? extends BasePojo> dao = DaoFactory.create(getDaoType());
+            dao.delete(id);
+            String message = "Entity with id = " + id + " deleted successfully";
+            return Response.ok(buildSuccess(message).toJSONString()).build();
+        } else {
+            return Response.status(Response.Status.FORBIDDEN).build();
         }
     }
 
-    void create(JSONObject json) {
-        BaseDao<? extends BasePojo> dao = DaoFactory.create(DaoType.REFERENCES);
-        dao.create(buildPojo(json));
+    void upload(SecurityContext context, InputStream fileInputStream, String fileName) {
+        if (context.isUserInRole("admin")) {
+            try (OutputStream out = new FileOutputStream(new File(getUploadPath().toString(), fileName))) {
+                int read;
+                byte[] bytes = new byte[1024];
+
+                while ((read = fileInputStream.read(bytes)) != -1) {
+                    out.write(bytes, 0, read);
+                }
+            } catch (IOException e) {
+                getLogger().error(getClass(), String.format("Error while uploading '%s' file", fileName));
+            }
+        }
+    }
+
+    Response create(SecurityContext context, JSONObject json) {
+        if (context.isUserInRole("admin")) {
+            BaseDao<? extends BasePojo> dao = DaoFactory.create(DaoType.REFERENCES);
+            dao.create(buildPojo(json));
+            String message = "Entity created successfully";
+            return Response.ok(buildSuccess(message).toJSONString()).build();
+        } else {
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
     }
 
     @SuppressWarnings("unchecked")
